+++ ca/django_ca/management/mixins.py
@@ -19,6 +19,7 @@ from textwrap import indent
 from typing import Optional
 
 from cryptography import x509
+from cryptography.hazmat.primitives.asymmetric.types import CertificateIssuerPrivateKeyTypes
 from cryptography.hazmat.primitives.serialization import Encoding
 
 from django.core.exceptions import ImproperlyConfigured
@@ -28,12 +29,7 @@ from django_ca import ca_settings
 from django_ca.extensions import extension_as_text, get_extension_name
 from django_ca.management import actions
 from django_ca.models import CertificateAuthority, X509CertMixin
-from django_ca.typehints import (
-    ActionsContainer,
-    AllowedHashTypes,
-    CertificateIssuerPrivateKeyTypes,
-    ParsableKeyType,
-)
+from django_ca.typehints import ActionsContainer, AllowedHashTypes, ParsableKeyType
 from django_ca.utils import add_colons, validate_public_key_parameters
 
 if typing.TYPE_CHECKING: