+++ ca/django_ca/tests/base/mixins.py
@@ -248,13 +248,7 @@ class TestCaseMixin(TestCaseProtocol):  # pylint: disable=too-many-public-method
             raise TypeError()  # just to make mypy happy
 
         self.assertIsInstance(parsed_crl.signature_hash_algorithm, type(algorithm))
-
-        # cryptography can not validate signatures of CRLs for non DSA/RSA/EC keys
-        # https://github.com/pyca/cryptography/issues/8156
-        if settings.CRYPTOGRAPHY_VERSION >= (40, 0) or not isinstance(  # pragma: cryptography<40.0 branch
-            public_key, (ed448.Ed448PublicKey, ed25519.Ed25519PublicKey)
-        ):
-            self.assertTrue(parsed_crl.is_signature_valid(public_key))
+        self.assertTrue(parsed_crl.is_signature_valid(public_key))
         self.assertEqual(parsed_crl.issuer, signer.pub.loaded.subject)
         self.assertEqual(parsed_crl.last_update, datetime.utcnow())
         self.assertEqual(parsed_crl.next_update, expires_timestamp)